# apis/swagger.yml
...
swagger: "2.0"
schemes:
  - "http"
  - "https"
produces:
  - "application/json"
  - "text/plain"
consumes:
  - "application/json"
  - "text/plain"
basePath: "/v1.24"
info:
  title: "Pouch Engine API"
  version: "1.24"
  description: |
    API is an HTTP API served by Pouch Engine.
paths:
...
  /info:
    get:
      summary: "Get System information"
      description: ""
      responses:
        200:
          schema:
            $ref: '#/definitions/SystemInfo'
          description: "no error"
        500:
          $ref: "#/responses/500ErrorResponse"
  /auth:
    post:
      summary: "Check auth configuration"
      description: "Validate credentials for a registry and, if available, get an identity token for accessing the registry without password."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "An identity token was generated successfully."
          schema:
            $ref: "#/definitions/AuthResponse"
        401:
          $ref: "#/responses/401ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "authConfig"
          in: "body"
          description: "Authentication to check"
          schema:
            $ref: "#/definitions/AuthConfig"
  /daemon/update:
    post:
      summary: "Update daemon's labels and image proxy"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
        400:
          description: "bad parameter"
          schema:
            $ref: '#/definitions/Error'
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "DaemonUpdateConfig"
          in: body
          description: "Config used to update daemon, only labels and image proxy are allowed."
          schema:
            $ref: "#/definitions/DaemonUpdateConfig"
  /events:
    get:
      summary: "Subscribe pouchd events to users"
      description: |
        Stream real-time events from the server.
        Report various object events of pouchd when something happens to them.
        Containers report these events: create`, `destroy`, `die`, `oom`, `pause`, `rename`, `resize`, `restart`, `start`, `stop`, `top`, `unpause`, `update` and `exec_die`
        Images report these events: `pull`, `untag`
        Volumes report these events: `create`, `destroy`
        Networks report these events: `create`, `connect`, `disconnect`, `destroy`
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
          schema:
            $ref: '#/definitions/EventsMessage'
        400:
          description: "bad parameter"
          schema:
            $ref: '#/definitions/Error'
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "since"
          in: "query"
          description: "Show events created since this timestamp then stream new events."
          type: "string"
        - name: "until"
          in: "query"
          description: "Show events created until this timestamp then stop streaming"
          type: "string"
        - name: "filters"
          in: "query"
          description: |
            A JSON encoded value of filters (a `map[string][]string`) to process on the event list. Available filters:
            - `container=<string>` container name or ID
            - `event=<string>` event type
            - `image=<string>` image name or ID
            - `label=<string>` image or container label
            - `network=<string>` network name or ID
            - `type=<string>` object to filter by, one of `container`, `image`, `volume`, `network`
            - `volume=<string>` volume name
          type: "string"
  /images/create:
    post:
      summary: "Create an image by pulling from a registry or importing from an existing source file"
      consumes:
        - "text/plain"
        - "application/octet-stream"
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
        404:
          schema:
            $ref: '#/definitions/Error'
          description: "image not found"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "fromImage"
          in: "query"
          description: "Name of the image to pull. The name may include a tag or digest. This parameter may only be used when pulling an image. The pull is cancelled if the HTTP connection is closed."
          type: "string"
        - name: "fromSrc"
          in: "query"
          description: "Source to import. The value may be a URL from which the image can be retrieved or `-` to read the image from the request body. This parameter may only be used when importing an image."
          type: "string"
        - name: "repo"
          in: "query"
          description: "Repository name given to an image when it is imported. The repo may include a tag. This parameter may only be used when importing an image."
          type: "string"
        - name: "tag"
          in: "query"
          description: "Tag or digest. If empty when pulling an image, this causes all tags for the given image to be pulled."
          type: "string"
        - name: "inputImage"
          in: "body"
          description: "Image content if the value `-` has been specified in fromSrc query parameter"
          schema:
            type: "string"
          required: false
        - name: "X-Registry-Auth"
          in: "header"
          description: "A base64-encoded auth configuration. [See the authentication section for details.](#section/Authentication)"
          type: "string"
  /images/load:
     post:
      summary: "Import images"
      description: |
        Load a set of images by oci.v1 format tar stream
      consumes:
        - application/x-tar
      responses:
        200:
          description: "no error"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "imageTarStream"
          in: "body"
          description: "tar stream containing images"
          schema:
            type: "string"
            format: "binary"
        - name: "name"
          in: "query"
          description: "set the image name for the tar stream, default unknown/unknown"
          type: "string"
  /images/save:
    get:
      summary: "Save image"
      description: |
        Save an image by oci.v1 format tar stream.
      produces:
        - application/x-tar
      responses:
        200:
          description: "no error"
          schema:
            type: "string"
            format: "binary"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "name"
          in: "query"
          description: "Image name which is to be saved"
          type: "string"
  /images/{imageid}/json:
    get:
      summary: "Inspect an image"
      description: "Return the information about image"
      operationId: "ImageInspect"
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
          schema:
            $ref: "#/definitions/ImageInfo"
          examples:
            application/json:
              CreatedAt: "2017-12-19 15:32:09"
              Digest: "sha256:e216a057b1cb1efc11f8a268f37ef62083e70b1b38323ba252e25ac88904a7e8"
              ID: "e216a057b1cb"
              Name: "ubuntu:12.04"
              Size: 103579269
              Tag: "12.04"
        404:
          description: "no such image"
          schema:
            $ref: "#/definitions/Error"
          examples:
            application/json:
              message: "No such image: e216a057b1cb"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - $ref: "#/parameters/imageid"
  /images/{imageid}/history:
    get:
      summary: "Get an image's history"
      description: "Return the history of each layer of image"
      operationId: "ImageHistory"
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HistoryResultItem"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - $ref: "#/parameters/imageid"
  /images/json:
    get:
      summary: "List Images"
      description: "Return a list of stored images."
      operationId: "ImageList"
      produces:
        - "application/json"
      responses:
        200:
          description: "Summary image data for the images matching the query"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ImageInfo"
          examples:
            application/json:
              - Id: "sha256:e216a057b1cb1efc11f8a268f37ef62083e70b1b38323ba252e25ac88904a7e8"
                ParentId: ""
                RepoTags:
                  - "ubuntu:12.04"
                  - "ubuntu:precise"
                RepoDigests:
                  - "ubuntu@sha256:992069aee4016783df6345315302fa59681aae51a8eeb2f889dea59290f21787"
                Created: 1474925151
                Size: 103579269
                VirtualSize: 103579269
                SharedSize: 0
                Labels: {}
                Containers: 2
              - Id: "sha256:3e314f95dcace0f5e4fd37b10862fe8398e3c60ed36600bc0ca5fda78b087175"
                ParentId: ""
                RepoTags:
                  - "ubuntu:12.10"
                  - "ubuntu:quantal"
                RepoDigests:
                  - "ubuntu@sha256:002fba3e3255af10be97ea26e476692a7ebed0bb074a9ab960b2e7a1526b15d7"
                  - "ubuntu@sha256:68ea0200f0b90df725d99d823905b04cf844f6039ef60c60bf3e019915017bd3"
                Created: 1403128455
                Size: 172064416
                VirtualSize: 172064416
                SharedSize: 0
                Labels: {}
                Containers: 5
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "all"
          in: "query"
          description: "Show all images. Only images from a final layer (no children) are shown by default."
          type: "boolean"
        - name: "filters"
          in: "query"
          description: |
            A JSON encoded value of the filters (a `map[string][]string`) to process on the images list. Available filters:

            - `before`=(`<image-name>[:<tag>]`,  `<image id>` or `<image@digest>`)
            - `reference`=(`<image-name>[:<tag>]`)
            - `since`=(`<image-name>[:<tag>]`,  `<image id>` or `<image@digest>`)
          type: "string"
        - name: "digests"
          in: "query"
          description: "Show digest information as a `RepoDigests` field on each image."
          type: "boolean"
  /images/search:
    get:
      summary: "Search images"
      produces:
        - "application/json"
      responses:
        200:
          description: "No error"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SearchResultItem"
        500:
          $ref: "#/responses/500ErrorResponse"
  /images/{imageid}/tag:
    post:
      summary: "Tag an image"
      description: "Add tag reference to the existing image"
      parameters:
        - $ref: "#/parameters/imageid"
        - name: "repo"
          in: "query"
          description: "The repository to tag in. For example, `someuser/someimage`."
          type: "string"
        - name: "tag"
          in: "query"
          description: "The name of the new tag."
          type: "string"
      responses:
        201:
          description: "No error"
        400:
          description: "Bad parameter"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "no such image"
          schema:
            $ref: "#/definitions/Error"
        500:
          $ref: "#/responses/500ErrorResponse"
  /images/{imageid}:
    delete:
      summary: "Remove an image"
      description: "Remove an image by reference."
      parameters:
        - $ref: "#/parameters/imageid"
        - name: "force"
          in: "query"
          description: "Remove the image even if it is being used"
          type: "boolean"
          default: false
      responses:
        204:
          description: "No error"
        404:
          description: "no such image"
          schema:
            $ref: "#/definitions/Error"
          examples:
            application/json:
              message: "No such image: c2ada9df5af8"
        500:
          $ref: "#/responses/500ErrorResponse"
  /containers/create:
    post:
      summary: "Create a container"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "name"
          in: "query"
          description: "Assign the specified name to the container. Must match `/?[a-zA-Z0-9_-]+`."
          type: "string"
          pattern: "/?[a-zA-Z0-9_-]+"
        - name: "body"
          in: "body"
          description: "Container to create"
          schema:
            $ref: "#/definitions/ContainerCreateConfig"
          required: true
      responses:
        201:
          description: "Container created successfully"
          schema:
            $ref: "#/definitions/ContainerCreateResp"
          examples:
            application/json:
              Id: "e90e34656806"
              Warnings: []
        400:
          description: "bad parameter"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "no such image"
          schema:
            $ref: "#/definitions/Error"
          examples:
            application/json:
              message: "image: xxx:latest: not found"
        409:
          description: "conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/json:
    get:
      summary: "Inspect a container"
      description: "Return low-level information about a container."
      operationId: "ContainerInspect"
      produces:
        - "application/json"
      responses:
        200:
          description: "no error"
          schema:
            $ref: "#/definitions/ContainerJSON"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - $ref: "#/parameters/id"
        - name: "size"
          in: "query"
          type: "boolean"
          description: "Return the size of container as fields `SizeRw` and `SizeRootFs`"
      tags: ["Container"]
  /containers/json:
    get:
      summary: "List containers"
      operationId: "ContainerList"
      produces: ["application/json"]
      responses:
        200:
          description: "Summary containers that matches the query"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Container"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - name: "all"
          in: "query"
          description: "Return all containers. By default, only running containers are shown"
          type: "boolean"
          default: false
        - name: "filters"
          in: "query"
          description: |
            Filters encoded as JSON string(type map[string][]string in Golang). This API will list containers match all of the filters. For example, `{"status": ["paused"]}` will only return paused containers.
            Available filters:
            - `id=<ID>` container ID filter, support regular expression.
            - `name=<name>` container name filter, support regular expression.
            - `status=<status>` container status filter, support regular expression.
            - `label=<key>=<value>` container label filter, support equal and unequal operator. such as `label=[k=a,k!=b]`.
          type: "string"
  /containers/{id}/rename:
    post:
      summary: "Rename a container"
      operationId: "ContainerRename"
      parameters:
        - $ref: "#/parameters/id"
        - name: "name"
          in: "query"
          required: true
          description: "New name for the container"
          type: "string"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        409:
          description: "name already in use"
          schema:
            $ref: "#/definitions/Error"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/restart:
    post:
      summary: "Restart a container"
      operationId: "ContainerRestart"
      parameters:
        - $ref: "#/parameters/id"
        - name: "name"
          in: "query"
          required: true
          description: "New name for the container"
          type: "string"
        - name: "t"
          in: "query"
          description: "Number of seconds to wait before killing the container"
          type: "integer"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/start:
    post:
      summary: "Start a container"
      operationId: "ContainerStart"
      parameters:
        - $ref: "#/parameters/id"
        - name: "detachKeys"
          in: "query"
          description: "Override the key sequence for detaching a container. Format is a single character `[a-Z]` or `ctrl-<value>` where `<value>` is one of: `a-z`, `@`, `^`, `[`, `,` or `_`."
          type: "string"
        - name: "checkpoint-dir"
          in: "query"
          description: "checkpoint directory"
          type: "string"
        - name: "checkpoint"
          in: "query"
          description: "checkpoint id"
          type: "string"
      responses:
        204:
          description: "no error"
        304:
          description: "container already started"
        404:
          $ref: "#/responses/404ErrorResponse"
        409:
          description: "container is paused"
          schema:
            $ref: "#/definitions/Error"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/stop:
    post:
      summary: "Stop a container"
      operationId: "ContainerStop"
      parameters:
        - $ref: "#/parameters/id"
        - name: "t"
          in: "query"
          description: "Number of seconds to wait before killing the container"
          type: "integer"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/pause:
    post:
      summary: "Pause a container"
      operationId: "ContainerPause"
      parameters:
        - $ref: "#/parameters/id"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/unpause:
    post:
      summary: "Unpause a container"
      operationId: "ContainerUnpause"
      parameters:
        - $ref: "#/parameters/id"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/top:
    get:
      summary: "Display the running processes of a container"
      operationId: "ContainerTop"
      parameters:
        - $ref: "#/parameters/id"
        - name: "ps_args"
          in: "query"
          description: "top arguments"
          type: "string"
      responses:
        200:
          description: "no error"
          schema:
            $ref: "#/definitions/ContainerProcessList"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/wait:
    post:
      summary: "Block until a container stops, then returns the exit code."
      operationId: "ContainerWait"
      parameters:
        - $ref: "#/parameters/id"
      responses:
        200:
          description: "The container has exited."
          schema:
            type: "object"
            required: [StatusCode]
            properties:
              StatusCode:
                description: "Exit code of the container"
                type: "integer"
                x-nullable: false
              Error:
                description: "The error message of waiting container"
                type: "string"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}:
    delete:
      summary: "Remove one container"
      operationId: "ContainerRemove"
      parameters:
        - $ref: "#/parameters/id"
        - name: "force"
          in: "query"
          description: "If the container is running, force query is used to kill it and remove it forcefully."
          type: "boolean"
      responses:
        204:
          description: "no error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
  /containers/{id}/exec:
    post:
      summary: "Create an exec instance"
      description: "Run a command inside a running container."
      operationId: "ContainerExec"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        201:
          description: "no error"
          schema:
            $ref: "#/definitions/ExecCreateResp"
        404:
          $ref: "#/responses/404ErrorResponse"
        409:
          description: "container is paused"
          schema:
            $ref: "#/definitions/Error"
        500:
          $ref: "#/responses/500ErrorResponse"
      parameters:
        - $ref: "#/parameters/id"
        - name: "body"
          in: "body"
          schema:
            $ref: "#/definitions/ExecCreateConfig"
          required: true
      tags: ["Exec"]
...
  /containers/{id}/resize:
    post:
      summary: "changes the size of the tty for a container"
      operationId: "ContainerResize"
      parameters:
        - $ref: "#/parameters/id"
        - name: "h"
          in: "query"
          description: "height of the tty"
          type: "string"
        - name: "w"
          in: "query"
          description: "width of the tty"
          type: "string"
      responses:
        200:
          description: "no error"
        400:
          description: "bad parameter"
          schema:
            $ref: "#/definitions/Error"
        404:
          $ref: "#/responses/404ErrorResponse"
        500:
          $ref: "#/responses/500ErrorResponse"
      tags: ["Container"]
...
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: string
...
  DaemonUpdateConfig:
    type: "object"
    properties:
      Labels:
        description: "Labels indentified the attributes of daemon"
        type: "array"
        items:
          type: "string"
        example: ["storage=ssd", "zone=hangzhou"]
      ImageProxy:
        description: "Image proxy used to pull image."
        type: "string"
...
  AuthConfig:
    type: "object"
    properties:
      Username:
        type: "string"
      Password:
        type: "string"
      Auth:
        type: "string"
      ServerAddress:
        type: "string"
      IdentityToken:
        type: "string"
        description: "IdentityToken is used to authenticate the user and get an access token for the registry"
      RegistryToken:
        type: "string"
        description: "RegistryToken is a bearer token to be sent to a registry"
  AuthResponse:
    description: "The response returned by login to a registry"
    type: "object"
    required: [Status]
    properties:
      Status:
        description: "The status of the authentication"
        type: "string"
        x-nullable: false
      IdentityToken:
        description: "An opaque token used to authenticate a user after a successful login"
        type: "string"
        x-nullable: false
...
responses:
  401ErrorResponse:
    description: An unexpected 401 error occurred.
    schema:
      $ref: "#/definitions/Error"
  404ErrorResponse:
    description: An unexpected 404 error occurred.
    schema:
      $ref: "#/definitions/Error"
  500ErrorResponse:
    description: An unexpected server error occurred.
    schema:
      $ref: "#/definitions/Error"